x/membership/types/message_enroll_test.go
package types

import (
	"testing"

	"github.com/cdbo/brain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgEnroll_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgEnroll
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgEnroll{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgEnroll{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
